Rename Autoyast to Autoinst to be consistent with yast2-network
