+++ test/lib/y2users/autoinst/hash_reader_test.rb
@@ -22,12 +22,12 @@
 require_relative "../test_helper"
 
 require "y2users/config"
-require "y2users/autoyast/hash_reader"
+require "y2users/autoinst/hash_reader"
 
 # defines exported users
 require_relative "../../../fixtures/users_export"
 
-describe Y2Users::Autoyast::HashReader do
+describe Y2Users::Autoinst::HashReader do
   subject { described_class.new(USERS_EXPORT) }
 
   describe "#read_to" do
